+++ src/youtube_scrap.py
@@ -1,6 +1,6 @@
 from src.utils import cprint, getHTTPResponse
-from src.video import Video
-from src.channel import Channel
+from src.youtube import YoutubeVideo
+from src.youtube import YoutubeChannel
 from src.db import Database
 
 def scrap_youtube_help(script_name, arg):
@@ -21,7 +21,7 @@ def scrap_video_w_url(url):
     $ python manage.py -runscrap -video "https://www.youtube.com/watch?v=GgjrAJQmMVA&ab_channel=RubiusZ" -add
     """
     html_content = getHTTPResponse(url, responseType = 'text')
-    return Video(html_content=html_content)
+    return YoutubeVideo(html_content=html_content)
 
 def scrap_video_w_id(id):
     """
@@ -29,7 +29,7 @@ def scrap_video_w_id(id):
     Ejemplo:
     $ python manage.py -runscrap -video 55O5Gnwbp14 -save_html
     """
-    return Video(id=id, en_html_save=False)
+    return YoutubeVideo(id=id, en_html_save=False)
 
 def scrap_channel_w_url(url):
     """
@@ -38,7 +38,7 @@ def scrap_channel_w_url(url):
     $ python manage.py -runscrap -channel "https://www.youtube.com/@elxokas" -save_html
     """
     html_content = getHTTPResponse(url, responseType = 'text')
-    return Channel(html_content=html_content)
+    return YoutubeChannel(html_content=html_content)
 
 def scrap_channel_w_id(id):
     """
@@ -47,7 +47,7 @@ def scrap_channel_w_id(id):
     $ python manage.py -runscrap -channel UC9c4rND1mNP-XrWQjbFxp8g -save_html -add
     $ python manage.py -runscrap -channel UCFzGNDHEZ5-7d5UXxfHUcRg -save_html -add
     """
-    return Channel(id=id)
+    return YoutubeChannel(id=id)
 
 def scrap_youtube():
     """
@@ -68,7 +68,7 @@ def scrap_youtube():
         # Obtengo la informacion para cada canal
         for channel_id in channel_ids:
             print(f'Fetching HTML content for channel {channel_id}')
-            channels.append( Channel(id=channel_id) )
+            channels.append( YoutubeChannel(id=channel_id) )
 
         # Proceso cada canal individualmente
         for channel in channels:
